+++ src/iter/italgos.h
@@ -24,7 +24,7 @@ struct iter_dump_s;
 
 #ifndef MD_IS_SET
 #define MD_BIT(x) (1ul << (x))
-#define MD_IS_SET(x, y)	((x) & MD_BIT(y))
+#define MD_IS_SET(x, y)	((unsigned long)(x) & MD_BIT(y))
 #define MD_CLEAR(x, y) ((x) & ~MD_BIT(y))
 #define MD_SET(x, y)	((x) | MD_BIT(y))
 #endif
